pyproject.toml
[project]
name = "honestcybereval"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "aioshutil",
    "aiorwlock",
    "PyYAML",
    "pydantic",
    "GitPython",
    "inspect-ai",
]

[project.optional-dependencies]
openai = ["openai"]
azure-ai = ["azure-ai-inference"]
groq = ["groq"]
anthropic = ["anthropic"]
all = [
    "honestcybereval[openai,azure-ai,groq,anthropic]",
]
[dependency-groups]
dev = [
    "pyright",
    "ruff",
    "pre-commit",
]
typestubs = [
    "types-PyYAML",
    "types-requests",
]
all = [
    {include-group = "dev"},
    {include-group = "typestubs"},
]

[build-system]
requires = ["setuptools", "wheel"]

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
include = ["src"]
reportMissingTypeStubs = true
deprecateTypingAliases = true
executionEnvironments = [{ root = "src" }]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = [
    "RUF",  # Ruff rules
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "PLE",  # Pylint error
    "PLC",  # Pylint convention
    "F",  # Pyflakes
    "B",  # flake8-bugbear
    "FA",  # flake8-future-annotations
    "EXE",  # flake8-executable
    "Q",  # flake8-quotes
    "I",  # isort
    "UP",  # pyupgrade
]
ignore=[
    "E501",  # line length, ignored for now
]
[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["PLC0414"]
